+++ src/common/buffer/Constants.ts
@@ -147,11 +147,4 @@ export const enum ExtFlags {
   VARIANT_OFFSET = 0xE0000000
 }
 
-export const enum UnderlineStyle {
-  NONE = 0,
-  SINGLE = 1,
-  DOUBLE = 2,
-  CURLY = 3,
-  DOTTED = 4,
-  DASHED = 5
-}
+export { UnderlineStyle } from '../Types';
