+++ src/common/public/UnicodeApi.ts
@@ -3,7 +3,7 @@
  * @license MIT
  */
 
-import { ICoreTerminal } from '../Types';
+import { ICoreTerminal } from '../CoreTerminal';
 import { IUnicodeHandling, IUnicodeVersionProvider } from '@xterm/xterm';
 
 export class UnicodeApi implements IUnicodeHandling {
